♻️이미지 처리 chat 버전 오버라이딩을 위해 인터페이스 추가
